requirements/base.txt
Django==2.0.1
psycopg2==2.7.3.2
Pillow==5.0.0
django-mptt==0.9.0
